@@ -3,24 +3,24 @@
 		"@oddbird/popover-polyfill": "^0.5.0"
 	},
 	"devDependencies": {
-		"@drop-in/graffiti": "^0.2.2",
-		"@sveltejs/adapter-auto": "^3.2.4",
-		"@sveltejs/kit": "^2.5.27",
-		"@sveltejs/package": "^2.3.5",
-		"@sveltejs/vite-plugin-svelte": "4.0.0-next.7",
+		"@drop-in/graffiti": "0.3.4",
+		"@sveltejs/adapter-auto": "^3.3.0",
+		"@sveltejs/kit": "^2.7.2",
+		"@sveltejs/package": "^2.3.6",
+		"@sveltejs/vite-plugin-svelte": "4.0.0",
 		"@types/eslint": "^9.6.1",
-		"eslint": "^9.12.0",
+		"eslint": "^9.13.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.44.1",
+		"eslint-plugin-svelte": "^2.46.0",
 		"globals": "^15.11.0",
 		"prettier": "^3.1.1",
 		"prettier-plugin-svelte": "^3.2.7",
 		"publint": "^0.2.11",
-		"svelte": "5.0.0-next.246",
-		"svelte-check": "^4.0.4",
+		"svelte": "5.1.0",
+		"svelte-check": "^4.0.5",
 		"typescript": "^5.6.3",
-		"typescript-eslint": "^8.8.1",
-		"vite": "^5.4.8"
+		"typescript-eslint": "^8.11.0",
+		"vite": "^5.4.10"
 	},
 	"engines": {
 		"node": ">20.11.1"
