@@ -1,4 +1,4 @@
-import { settings } from '$settings';
+import settings from 'virtual:dropin-config';
 import { goto } from '$app/navigation';
 import { toaster } from '@drop-in/decks';
 
@@ -15,7 +15,7 @@ export function auth_form_state() {
 		error_message = e_message;
 	}
 
-	function success(route: string | boolean = settings.app_route) {
+	function success(route: string | boolean = settings.app.route) {
 		error_message = undefined;
 		status = 'SUCCESS';
 		if (route && typeof route === 'string') goto(route);
