@@ -1,7 +1,7 @@
 <script lang="ts">
-	import { settings } from '$settings';
+	import settings from 'virtual:dropin-config';
 </script>
 
-<h1>{settings.app_name}</h1>
+<h1>{settings.app.name}</h1>
 
 <a href="/auth/signup">Sign Up</a>
