@@ -1,17 +1,16 @@
 <script lang="ts">
-	import { create_zero } from '$/lib/z';
-	import Header from '$/routes/(app)/Header.svelte';
+	import { create_zero } from '$lib/z';
+	import Header from '$routes/(app)/Header.svelte';
 	import Footer from '$routes/(app)/Footer.svelte';
+	import { get_login } from '@drop-in/pass/client';
+	import { auth_guard } from '$utils/auth_guard';
 	// import Verify from '$/lib/auth/Verify.svelte';
 	let { children } = $props()
 	
-	const auth = get_login()
-	create_zero({
-		user_id: auth.user_id,
-		jwt: auth.jwt,
+
+	$effect.pre(() => {
+		auth_guard()
 	});
-
-
 	// TODO Auth Gaurd?
 </script>
 
