@@ -2,12 +2,13 @@
 	import Header from '$routes/(app)/Header.svelte';
 	import Footer from '$routes/(app)/Footer.svelte';
 	import { PUBLIC_SERVER } from '$env/static/public';
+	import { app_guard } from '$utils/app_guard';
 	// import { app_guard } from '$utils/app_guard';
 
 	let { children } = $props()
  
 	// Everything in the app route tree is protected behind user accounts.
-	// $effect.pre(app_guard);
+	$effect.pre(app_guard);
 </script>
 
 <!-- Delete me once you set your Pocketbase URL var -->
