@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { settings } from '$settings';
+	import settings from 'virtual:dropin-config';
 </script>
 
 <h1>{settings.app_name}</h1>
