@@ -1,5 +1,5 @@
 <script lang="ts">
-	import ForgotPassword from '$/lib/auth/ForgotPassword.svelte';
+	import ForgotPassword from '$lib/auth/ForgotPassword.svelte';
 </script>
 
 <ForgotPassword />
