@@ -1,8 +1,16 @@
 <script lang="ts">
 	import "@drop-in/graffiti";
 	import { Toast } from '@drop-in/decks';
+	import { get_login } from "@drop-in/pass/client";
+	import { create_zero } from "$lib/z";
 	let { children } = $props()
-
+	$effect.pre(() => {
+		const auth = get_login();
+		create_zero({
+			user_id: auth.sub || 'anon',
+			auth: auth.jwt,
+		});
+	});
 </script>
 
 <!-- If you would like to use the same layout in app and site, use this file -->
